# docs/openapi.snap.json
{"components":{"schemas":{"APIError":{"properties":{"code":{"$ref":"#/components/schemas/ErrorCode"},"details":{"additionalProperties":true,"default":{},"title":"Details","type":"object"},"message":{"title":"Message","type":"string"}},"required":["code","message"],"title":"APIError","type":"object"},"BudgetSpec":{"additionalProperties":false,"properties":{"max_cost":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Max Cost"},"max_generations":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Generations"},"max_rollouts":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Rollouts"}},"title":"BudgetSpec","type":"object"},"DatasetSpec":{"additionalProperties":false,"properties":{"name":{"title":"Name","type":"string"},"split":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Split"}},"required":["name"],"title":"DatasetSpec","type":"object"},"ErrorCode":{"enum":["unauthorized","not_found","rate_limited","payload_too_large","not_cancelable","sse_backpressure","validation_error","internal_error"],"title":"ErrorCode","type":"string"},"ErrorResponse":{"properties":{"error":{"$ref":"#/components/schemas/APIError"}},"required":["error"],"title":"ErrorResponse","type":"object"},"EvalStartRequest":{"properties":{"early_stop_patience":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Early Stop Patience"},"max_examples":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Examples"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"recombination_rate":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Recombination Rate"},"seed":{"anyOf":[{"type":"integer"},{"type":"null"}],"default":42,"title":"Seed"},"target_model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Model"},"tournament_size":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tournament Size"}},"title":"EvalStartRequest","type":"object"},"ExampleIn":{"additionalProperties":false,"properties":{"expected":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Expected"},"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"input":{"title":"Input","type":"string"},"meta":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Meta"}},"required":["input"],"title":"ExampleIn","type":"object"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail","type":"array"}},"title":"HTTPValidationError","type":"object"},"JobState":{"examples":[{"created_at":0.0,"job_id":"123e4567","result":{"proposal":"..."},"status":"finished","updated_at":1.0}],"properties":{"created_at":{"title":"Created At","type":"number"},"job_id":{"title":"Job Id","type":"string"},"result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Result"},"status":{"title":"Status","type":"string"},"updated_at":{"title":"Updated At","type":"number"}},"required":["job_id","status","created_at","updated_at"],"title":"JobState","type":"object"},"OptimizeRequest":{"additionalProperties":false,"examples":[{"early_stop_patience":3,"evaluation_rubric":"clarity, brevity, imagery","examples":[{"expected":"short","input":"long text"}],"objectives":["brevity","diversity","coverage"],"prompt":"Write a haiku","recombination_rate":0.5,"target_model_id":"gpt-4o-mini","tournament_size":4}],"properties":{"budget":{"anyOf":[{"$ref":"#/components/schemas/BudgetSpec"},{"type":"null"}]},"context":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Context"},"dataset":{"anyOf":[{"$ref":"#/components/schemas/DatasetSpec"},{"type":"null"}]},"early_stop_patience":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Early Stop Patience"},"evaluation_rubric":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Evaluation Rubric"},"examples":{"anyOf":[{"items":{"$ref":"#/components/schemas/ExampleIn"},"type":"array"},{"type":"null"}],"title":"Examples"},"max_tokens":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Tokens"},"metrics":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Metrics"},"mode":{"default":"default","enum":["default","gepa"],"title":"Mode","type":"string"},"objectives":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Objectives"},"prompt":{"title":"Prompt","type":"string"},"recombination_rate":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Recombination Rate"},"seed":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Seed"},"target_model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Model"},"temperature":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Temperature"},"tournament_size":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tournament Size"}},"required":["prompt"],"title":"OptimizeRequest","type":"object"},"OptimizeResponse":{"examples":[{"job_id":"123e4567"}],"properties":{"job_id":{"title":"Job Id","type":"string"}},"required":["job_id"],"title":"OptimizeResponse","type":"object"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location","type":"array"},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}},"required":["loc","msg","type"],"title":"ValidationError","type":"object"}}},"info":{"title":"gepa-next","version":"0.1.0"},"openapi":"3.1.0","paths":{"/v1/admin/jobs":{"get":{"operationId":"list_jobs_v1_admin_jobs_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response List Jobs V1 Admin Jobs Get","type":"object"}}},"description":"Successful Response"}},"summary":"List Jobs","tags":["admin"]}},"/v1/admin/jobs/{job_id}":{"delete":{"operationId":"delete_job_v1_admin_jobs__job_id__delete","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}}],"responses":{"204":{"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Delete Job","tags":["admin"]},"get":{"operationId":"get_job_v1_admin_jobs__job_id__get","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobState"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Job","tags":["admin"]}},"/v1/admin/jobs/{job_id}/cancel":{"post":{"operationId":"cancel_job_v1_admin_jobs__job_id__cancel_post","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobState"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"409":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Conflict"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Cancel Job","tags":["admin"]}},"/v1/eval/start":{"post":{"operationId":"eval_start_v1_eval_start_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvalStartRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OptimizeResponse"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Start evaluation job","tags":["eval"]}},"/v1/eval/{job_id}/events":{"get":{"description":"Server-Sent Events stream for evaluation jobs.","operationId":"eval_events_v1_eval__job_id__events_get","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"format":"uuid","title":"Job Id","type":"string"}}],"responses":{"200":{"content":{"text/event-stream":{}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Unprocessable Entity"}},"summary":"Stream eval events","tags":["eval"]}},"/v1/examples":{"get":{"operationId":"examples_list_v1_examples_get","parameters":[{"in":"query","name":"limit","required":false,"schema":{"default":50,"title":"Limit","type":"integer"}},{"in":"query","name":"offset","required":false,"schema":{"default":0,"title":"Offset","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response Examples List V1 Examples Get","type":"object"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Examples List","tags":["examples"]}},"/v1/examples/bulk":{"post":{"operationId":"examples_bulk_v1_examples_bulk_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ExampleIn"},"title":"Items","type":"array"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response Examples Bulk V1 Examples Bulk Post","type":"object"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Examples Bulk","tags":["examples"]}},"/v1/examples/{example_id}":{"delete":{"operationId":"examples_delete_v1_examples__example_id__delete","parameters":[{"in":"path","name":"example_id","required":true,"schema":{"title":"Example Id","type":"string"}}],"responses":{"204":{"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Examples Delete","tags":["examples"]}},"/v1/healthz":{"get":{"operationId":"healthz_v1_healthz_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"title":"Response Healthz V1 Healthz Get","type":"object"}}},"description":"Successful Response"}},"summary":"Healthz","tags":["v1"]}},"/v1/metrics":{"get":{"description":"Exposes Prometheus metrics in text format.","operationId":"metrics_prom_v1_metrics_get","responses":{"200":{"content":{"text/plain":{"schema":{"description":"Prometheus exposition","type":"string"}}},"description":"Prometheus metrics"}},"summary":"Prometheus metrics","tags":["v1"]}},"/v1/metricsz":{"get":{"operationId":"metricsz_v1_metricsz_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response Metricsz V1 Metricsz Get","type":"object"}}},"description":"Successful Response"}},"summary":"Metricsz","tags":["v1"]}},"/v1/optimize":{"post":{"description":"Create an optimization job. Use optional Idempotency-Key header to dedupe submissions.","operationId":"create_optimize_job_v1_optimize_post","parameters":[{"in":"query","name":"iterations","required":false,"schema":{"default":1,"title":"Iterations","type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OptimizeRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OptimizeResponse"}}},"description":"Successful Response"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Unauthorized"},"413":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Request Entity Too Large"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"},"429":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Too Many Requests"}},"summary":"Create optimization job","tags":["v1"]}},"/v1/optimize/{job_id}":{"delete":{"operationId":"cancel_job_endpoint_v1_optimize__job_id__delete","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobState"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"409":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Conflict"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Cancel Job Endpoint","tags":["v1"]},"get":{"operationId":"get_job_v1_optimize__job_id__get","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobState"}}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Job","tags":["v1"]}},"/v1/optimize/{job_id}/events":{"get":{"description":"Server-Sent Events stream. Use Last-Event-ID header to resume from a specific event id.","operationId":"optimize_events_v1_optimize__job_id__events_get","parameters":[{"in":"path","name":"job_id","required":true,"schema":{"title":"Job Id","type":"string"}}],"responses":{"200":{"content":{"text/event-stream":{}},"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Stream job events","tags":["v1"]}},"/v1/readyz":{"get":{"operationId":"readyz_v1_readyz_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"title":"Response Readyz V1 Readyz Get","type":"object"}}},"description":"Successful Response"}},"summary":"Readyz","tags":["v1"]}}}}